requirements.txt
pandas
boto3
python-dotenv
tenacity
scikit-learn
joblib
ipykernel
snowflake-connector-python
snowflake